# Conflicts:
#	midExam/SoftwarePlatform/src/main/java/com/qg24/softwareplatform/mapper/SoftwareMapper.java
#	midExam/SoftwarePlatform/src/main/resources/application.yml

+++ midExam/SoftwarePlatform/src/main/java/com/qg24/softwareplatform/po/entity/Software.java
@@ -17,7 +17,7 @@ public class Software {
 
     private String description;
 
-    private int downloadCount;
+    private double downloadCount;
 
     private int softwareId;
